Feat/split player data from db

+++ ArchaicQuestII.GameLogic/Account/AccountStats.cs
@@ -10,7 +10,7 @@ namespace ArchaicQuestII.GameLogic.Account
         public int MobKills { get; set; }
         public int PlayerKills { get; set; }
         public int Deaths { get; set; }
-        public int TotalPlayTime { get; set; }
+        public double TotalPlayTime { get; set; }
         public int ExploredRooms { get; set; }
     }
 }
